pyproject.toml
[tool.poetry]
name = "ai-character-backend"
version = "0.1.0"
description = ""
authors = ["Mogami <atsuyatsuchikawa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.3"
uvicorn = "^0.32.0"
sqlalchemy = "^2.0.36"
python-dotenv = "^1.0.1"
openai = "^1.52.2"
requests = "^2.32.3"
google-generativeai = "^0.8.3"
python-multipart = "^0.0.16"
qdrant-client = "^1.12.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
